feat(db): versioned schema migrations
Add schema_version tracking and apply_migrations on connect. Legacy
databases are stamped at v1 without data loss.
