+++ src/squidc5/db/store.py
@@ -11,187 +11,10 @@ from typing import Any
 
 import aiosqlite
 
-SCHEMA = """
-CREATE TABLE IF NOT EXISTS tokens (
-    id TEXT PRIMARY KEY,
-    name TEXT NOT NULL,
-    token_hash TEXT NOT NULL UNIQUE,
-    scopes TEXT NOT NULL,
-    mcp_tools TEXT NOT NULL DEFAULT '[]',
-    created_at REAL NOT NULL,
-    created_by TEXT,
-    expires_at REAL,
-    revoked INTEGER NOT NULL DEFAULT 0,
-    last_used_at REAL
-);
+from squidc5.db.migrate import BASELINE_SQL, apply_migrations
 
-CREATE TABLE IF NOT EXISTS sessions (
-    id TEXT PRIMARY KEY,
-    kind TEXT NOT NULL,
-    remote_addr TEXT,
-    user_agent TEXT,
-    hostname TEXT,
-    username TEXT,
-    os_info TEXT,
-    status TEXT NOT NULL DEFAULT 'active',
-    listener_id TEXT,
-    metadata TEXT NOT NULL DEFAULT '{}',
-    created_at REAL NOT NULL,
-    last_seen_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS listeners (
-    id TEXT PRIMARY KEY,
-    name TEXT NOT NULL,
-    kind TEXT NOT NULL,
-    host TEXT NOT NULL DEFAULT '0.0.0.0',
-    port INTEGER NOT NULL,
-    status TEXT NOT NULL DEFAULT 'stopped',
-    config TEXT NOT NULL DEFAULT '{}',
-    created_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS tasks (
-    id TEXT PRIMARY KEY,
-    session_id TEXT NOT NULL,
-    command TEXT NOT NULL,
-    args TEXT NOT NULL DEFAULT '{}',
-    status TEXT NOT NULL DEFAULT 'pending',
-    result TEXT,
-    created_by TEXT,
-    created_at REAL NOT NULL,
-    completed_at REAL,
-    FOREIGN KEY(session_id) REFERENCES sessions(id)
-);
-
-CREATE TABLE IF NOT EXISTS audit_log (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    ts REAL NOT NULL,
-    actor TEXT NOT NULL,
-    actor_type TEXT NOT NULL,
-    action TEXT NOT NULL,
-    resource TEXT,
-    details TEXT NOT NULL DEFAULT '{}',
-    risk_score INTEGER NOT NULL DEFAULT 0,
-    allowed INTEGER NOT NULL DEFAULT 1
-);
-
-CREATE TABLE IF NOT EXISTS llm_connections (
-    id TEXT PRIMARY KEY,
-    name TEXT NOT NULL,
-    provider TEXT NOT NULL,
-    base_url TEXT,
-    model TEXT NOT NULL,
-    api_key_enc TEXT,
-    capabilities TEXT NOT NULL DEFAULT '[]',
-    enabled INTEGER NOT NULL DEFAULT 1,
-    created_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS policies (
-    id TEXT PRIMARY KEY,
-    name TEXT NOT NULL UNIQUE,
-    rules TEXT NOT NULL,
-    enabled INTEGER NOT NULL DEFAULT 1,
-    created_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS metrics (
-    key TEXT PRIMARY KEY,
-    value REAL NOT NULL,
-    updated_at REAL NOT NULL
-);
-
-CREATE INDEX IF NOT EXISTS idx_audit_ts ON audit_log(ts);
-CREATE INDEX IF NOT EXISTS idx_tasks_session ON tasks(session_id);
-CREATE INDEX IF NOT EXISTS idx_sessions_status ON sessions(status);
-
-CREATE TABLE IF NOT EXISTS c2_profiles (
-    id TEXT PRIMARY KEY,
-    name TEXT NOT NULL,
-    config TEXT NOT NULL,
-    active INTEGER NOT NULL DEFAULT 0,
-    created_at REAL NOT NULL,
-    updated_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS teams (
-    id TEXT PRIMARY KEY,
-    name TEXT NOT NULL UNIQUE,
-    created_by TEXT,
-    created_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS session_handoffs (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    session_id TEXT NOT NULL,
-    payload TEXT NOT NULL,
-    ts REAL NOT NULL
-);
-
-CREATE INDEX IF NOT EXISTS idx_handoffs_session ON session_handoffs(session_id);
-
-CREATE TABLE IF NOT EXISTS operator_chat (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    team_id TEXT,
-    actor TEXT NOT NULL,
-    message TEXT NOT NULL,
-    ts REAL NOT NULL
-);
-
-CREATE INDEX IF NOT EXISTS idx_chat_ts ON operator_chat(ts);
-
-CREATE TABLE IF NOT EXISTS plugins (
-    name TEXT PRIMARY KEY,
-    version TEXT NOT NULL,
-    manifest TEXT NOT NULL,
-    signature TEXT NOT NULL,
-    enabled INTEGER NOT NULL DEFAULT 0,
-    created_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS operator_notes (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    session_id TEXT NOT NULL,
-    actor TEXT NOT NULL,
-    note TEXT NOT NULL,
-    ts REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS team_members (
-    team_id TEXT NOT NULL,
-    actor TEXT NOT NULL,
-    role TEXT NOT NULL DEFAULT 'operator',
-    added_at REAL NOT NULL,
-    PRIMARY KEY (team_id, actor)
-);
-
-CREATE TABLE IF NOT EXISTS oast_clients (
-    id TEXT PRIMARY KEY,
-    token TEXT NOT NULL UNIQUE,
-    label TEXT NOT NULL DEFAULT '',
-    created_by TEXT,
-    meta TEXT NOT NULL DEFAULT '{}',
-    created_at REAL NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS oast_interactions (
-    id TEXT PRIMARY KEY,
-    client_id TEXT,
-    token TEXT,
-    protocol TEXT NOT NULL,
-    listener_id TEXT,
-    remote TEXT,
-    raw TEXT NOT NULL DEFAULT '{}',
-    correlation_key TEXT,
-    ts REAL NOT NULL
-);
-
-CREATE INDEX IF NOT EXISTS idx_oast_clients_token ON oast_clients(token);
-CREATE INDEX IF NOT EXISTS idx_oast_interactions_ts ON oast_interactions(ts);
-CREATE INDEX IF NOT EXISTS idx_oast_interactions_client ON oast_interactions(client_id);
-CREATE INDEX IF NOT EXISTS idx_oast_interactions_token ON oast_interactions(token);
-"""
+# Back-compat alias for anything importing SCHEMA
+SCHEMA = BASELINE_SQL
 
 def _now() -> float:
     return time.time()
@@ -215,7 +38,7 @@ class Database:
         self._db.row_factory = aiosqlite.Row
         await self._db.execute("PRAGMA journal_mode=WAL")
         await self._db.execute("PRAGMA foreign_keys=ON")
-        await self._db.executescript(SCHEMA)
+        await apply_migrations(self._db)
         await self._db.commit()
 
     async def close(self) -> None:
